Created - economics.py with economics functions - misc with some plotting functions all the example files have been tested and they are running.
